🔭 Stellar: Complete TeleVue audit and ZWO ghost cleanup
- Updated TeleVue telescope database with verified specifications (aperture,
  focal length, mass) for all 8 models (TV-60, TV-76, TV-85, TV-102,
  NP101, NP101is, NP127is, NP127fli).
- Pruned 12 non-existent 'ghost' mono ZWO models incorrectly based on
  color-only sensors (ASI071, ASI094, ASI128, ASI2400, ASI482, ASI485,
  ASI664, ASI676, ASI715).
- Fixed regressions in existing test suites (test_stellar_zwo_cameras.py,
  test_stellar_v6.py, test_stellar_v7.py) caused by ghost model removal.
- Added comprehensive TeleVue verification tests (test_televue_specs.py).
- Verified database integrity and formatting for vendor modules using AST pruning.
