+++ apts/opticalequipment/filter_wheel/vendors/ts_optics.py
@@ -1,0 +1,40 @@
+from ..base import FilterHolder
+
+
+class TsOpticsFilterHolder(FilterHolder):
+    _DATABASE = {
+        "TS_Optics_Filter_Drawer_M48": {
+            "brand": "TS-Optics",
+            "name": "Filter Drawer (M48)",
+            "type": "type_filter_holder",
+            "optical_length": 25,
+            "mass": 200,
+            "tside_thread": "M48",
+            "tside_gender": "Male",
+            "cside_thread": "M48",
+            "cside_gender": "Female",
+            "reversible": False,
+            "bf_role": "",
+        },
+        "TS_Optics_Filter_Drawer_M54": {
+            "brand": "TS-Optics",
+            "name": "Filter Drawer (M54)",
+            "type": "type_filter_holder",
+            "optical_length": 25,
+            "mass": 230,
+            "tside_thread": "M54",
+            "tside_gender": "Male",
+            "cside_thread": "M54",
+            "cside_gender": "Female",
+            "reversible": False,
+            "bf_role": "",
+        },
+    }
+
+    @classmethod
+    def TS_Optics_Filter_Drawer_M48(cls):
+        return cls.from_database(cls._DATABASE["TS_Optics_Filter_Drawer_M48"])
+
+    @classmethod
+    def TS_Optics_Filter_Drawer_M54(cls):
+        return cls.from_database(cls._DATABASE["TS_Optics_Filter_Drawer_M54"])
